+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   - :beetle: Fix geometry error on delete
   - :beetle: Fix non existing geometry in simulation/geometry
   - :arrow_up: Remove ThreeJS (moved to tanatloc-3d)
+- Plugins:
+  - :beetle: Fix Rescale plugin
+  - :beetle: Fix Sharetask plugin
 
 ---